refactor(phrases): split I18n locale file
