+++ packages/phrases/src/language/en/translation/admin-console/api-resource-details.ts
