+++ packages/phrases/src/language/en/translation/admin-console/api-resources.ts
