+++ packages/phrases/src/language/en/translation/admin-console/application-details.ts
