+++ packages/phrases/src/language/en/translation/admin-console/applications.ts
