+++ packages/phrases/src/language/en/translation/admin-console/connector-details.ts
