+++ packages/phrases/src/language/en/translation/admin-console/connectors.ts
