+++ packages/phrases/src/language/en/translation/admin-console/dashboard.ts
