+++ packages/phrases/src/language/en/translation/admin-console/get-started.ts
