+++ packages/phrases/src/language/en/translation/admin-console/log-details.ts
