+++ packages/phrases/src/language/en/translation/admin-console/tab-sections.ts
