+++ packages/phrases/src/language/en/translation/admin-console/tabs.ts
