+++ packages/phrases/src/language/en/translation/admin-console/user-details.ts
