+++ packages/phrases/src/language/en/translation/admin-console/welcome.ts
