+++ packages/phrases/src/language/zh-cn/translation/admin-console/api-resource-details.ts
