+++ packages/phrases/src/language/zh-cn/translation/admin-console/api-resources.ts
