+++ packages/phrases/src/language/zh-cn/translation/admin-console/application-details.ts
