+++ packages/phrases/src/language/zh-cn/translation/admin-console/applications.ts
