+++ packages/phrases/src/language/zh-cn/translation/admin-console/connector-details.ts
