+++ packages/phrases/src/language/zh-cn/translation/admin-console/dashboard.ts
