+++ packages/phrases/src/language/zh-cn/translation/admin-console/errors.ts
