+++ packages/phrases/src/language/zh-cn/translation/admin-console/get-started.ts
