+++ packages/phrases/src/language/zh-cn/translation/admin-console/log-details.ts
