+++ packages/phrases/src/language/zh-cn/translation/admin-console/tab-sections.ts
