+++ packages/phrases/src/language/zh-cn/translation/admin-console/tabs.ts
