+++ packages/phrases/src/language/zh-cn/translation/admin-console/user-details.ts
