+++ packages/phrases/src/language/zh-cn/translation/demo-app.ts
