+++ packages/phrases/src/language/zh-cn/translation/index.ts
