+++ packages/phrases/src/locales/en.ts
@@ -1,3 +1,5 @@
-// Note: we need to re-export the locale under this `locales` folder,
-// so that the `useTranslation` function can infer the i18n key type.
-export { default } from '../language-pack/en';
+/**
+ * Note: we need to re-export the locale under this `locales` folder,
+ * so that the `useTranslation` function can infer the i18n key type.
+ */
+export { default } from '../language/en';
