+++ packages/phrases/src/locales/zh-cn.ts
@@ -1,3 +1,5 @@
-// Note: we need to re-export the locale under this `locales` folder,
-// so that the `useTranslation` function can infer the i18n key type.
-export { default } from '../language-pack/zh-cn';
+/**
+ * Note: we need to re-export the locale under this `locales` folder,
+ * so that the `useTranslation` function can infer the i18n key type.
+ */
+export { default } from '../language/zh-cn';
